refactor(engine-http): move content connection routing out of the controller
The controller closure had grown to ~275 lines and thirteen responsibilities, of
which the read-after-write routing tree was the newest and the least related to
HTTP. ContentConnectionRouter now owns it: the connection cap, the routing
decision, the pinned scope with its probe and disposal, and the three fallbacks.

The controller keeps what is genuinely its own - resolving the request and turning
the outcome into response headers - and hands the router a `run` callback. The
router takes the connections it may choose between rather than the whole
ProjectContainer, so it has no dependency on the container, on Koa, or on the
request; nothing about it needed to become generic, since the response type is the
only one it will ever carry.

Behaviour is unchanged and the existing routing suite covers it as written: 325
lines of controller become 248, plus a 158-line unit with a name.
